# contracts/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Data Structure Agent API
  version: 1.0.0
  description: Spring backend contract for the data structure learning agent.
servers:
  - url: /api/v1
    description: Relative base; usable from structify.cn and admin.structify.cn. Admin routes still require a database-backed ADMIN role.
security: []
...
paths:
  /auth/request-code:
    post:
      tags: [Auth]
      summary: Request an email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestCodeRequest'
      responses:
        '200':
          description: Generic delivery response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationCodeDelivery'
        '400': {$ref: '#/components/responses/BadRequest'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a verified account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Authenticated account; also sets the HttpOnly session cookie
          headers:
            Set-Cookie:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email or username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated account; also sets the HttpOnly session cookie
          headers:
            Set-Cookie:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset a password with a reset-purpose verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password updated and a new session issued
          headers:
            Set-Cookie:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /learning/progress:
    get:
      tags: [Learning]
      summary: Aggregate owned learning activity by chapter
      security: [{cookieAuth: []}, {bearerAuth: []}]
      responses:
        '200':
          description: Explainable activity counts, not an opaque score
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningProgress'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /learning/workbench:
    get:
      tags: [Learning]
      summary: Restore the caller's latest valid persisted algorithm scene
      description: >-
        Returns aggregate owned activity and, when present, the caller's latest
        renderer-compatible DSVP scene for a published chapter. A missing or malformed
        historical payload is omitted instead of being represented as a saved scene.
      security: [{cookieAuth: []}, {bearerAuth: []}]
      responses:
        '200':
          description: Owned learning-workbench context
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningWorkbench'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /learning/events:
    post:
      tags: [Learning]
      summary: Record an authenticated student's reviewed learning activity
      security: [{cookieAuth: []}, {bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LearningEventRequest'}
      responses:
        '201':
          description: Learning activity recorded
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningEvent'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
...
  /admin/sandbox-config:
    get:
      tags: [Admin]
      summary: Read the configured code execution sandbox capability
      description: >-
        Returns a credential-free view of the persisted sandbox target. A disabled
        or missing configuration is reported as unavailable and execution remains
        fail-closed. Supported providers are PISTON and JUDGE0.
      security: [{cookieAuth: []}, {bearerAuth: []}]
      responses:
        '200':
          description: Sandbox capability and safe configuration view
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SandboxConfigCapability'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    put:
      tags: [Admin]
      summary: Save and enable or disable the code execution sandbox
      description: >-
        Saves a validated HTTPS service URL (local HTTP is allowed only in dev/test),
        resets prior connection evidence, and records an administrative audit event.
      security: [{cookieAuth: []}, {bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateSandboxConfigRequest'}
      responses:
        '200':
          description: Safe persisted sandbox configuration view
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SandboxConfig'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/sandbox-config/test:
    post:
      tags: [Admin]
      summary: Probe the saved sandbox connection
      security: [{cookieAuth: []}, {bearerAuth: []}]
      responses:
        '200':
          description: Non-sensitive connection result
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SandboxConnectionTest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ds_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Unauthorized:
      description: Authentication required or credentials invalid
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Forbidden:
      description: Authenticated user cannot access the requested source
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
    Conflict:
      description: State conflict
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    PayloadTooLarge:
      description: Input exceeds configured limit
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    TooManyRequests:
      description: Rate limit exceeded
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
  schemas:
    ApiError:
      type: object
      required: [code, message, requestId, details]
      properties:
        code: {type: string}
        message: {type: string}
        requestId: {type: string}
        details:
          type: array
          items: {type: string}
    RequestCodeRequest:
      type: object
      required: [email, purpose]
      properties:
        email: {type: string, format: email}
        purpose: {type: string, enum: [register, reset]}
    VerificationCodeDelivery:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Generic response. The verification code is delivered out of band and is never returned by HTTP.
    RegisterRequest:
      type: object
      required: [email, code, password]
      properties:
        email: {type: string, format: email}
        username:
          type: string
          nullable: true
          pattern: '^[A-Za-z0-9_]{3,32}$'
          description: Optional public account identifier. Case is preserved for display and compared case-insensitively.
        code: {type: string, minLength: 6, maxLength: 6}
        password: {type: string, minLength: 8, maxLength: 128}
    LoginRequest:
      type: object
      required: [password]
      oneOf:
        - required: [email]
        - required: [username]
      properties:
        email:
          type: string
          maxLength: 254
          description: Backward-compatible login identifier field. It accepts either an email address or a username.
        username:
          type: string
          pattern: '^[A-Za-z0-9_]{3,32}$'
          description: Explicit username login field. Supply this or email, not both.
        password: {type: string, maxLength: 128}
    ResetPasswordRequest:
      type: object
      required: [email, code, password]
      properties:
        email: {type: string, format: email}
        code: {type: string, minLength: 6, maxLength: 6}
        password: {type: string, minLength: 8, maxLength: 128}
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: {type: string}
        user: {$ref: '#/components/schemas/User'}
    User:
      type: object
      required: [id, email, roles]
      properties:
        id: {type: integer, format: int64}
        email: {type: string, format: email}
        username:
          type: string
          nullable: true
          pattern: '^[A-Za-z0-9_]{3,32}$'
        roles:
          type: array
          items: {type: string, enum: [STUDENT, TEACHER, ADMIN]}
...
    AnimationStep:
      type: object
      required: [op, label, note]
      properties:
        op: {type: string, maxLength: 32}
        label: {type: string, maxLength: 48}
        note: {type: string, maxLength: 240}
        value:
          oneOf:
            - {type: string, maxLength: 48}
            - {type: number}
            - {type: boolean}
            - {type: object, nullable: true, enum: [null]}
        index: {type: integer, minimum: 0, maximum: 1024, nullable: true}
        node: {type: integer, minimum: 0, maximum: 64, nullable: true}
        i: {type: integer, minimum: 0, maximum: 1024, nullable: true}
        j: {type: integer, minimum: 0, maximum: 1024, nullable: true}
        key: {type: string, maxLength: 48, nullable: true}
        val: {type: string, maxLength: 48, nullable: true}
    AnimationDefinition:
      type: object
      required: [animation, type, title, description, initial, steps]
      properties:
        animation: {type: boolean, enum: [true]}
        type: {type: string, enum: [stack, list, tree, queue, heap, hash, array]}
        title: {type: string, maxLength: 60}
        description: {type: string, maxLength: 240}
        initial:
          type: array
          maxItems: 64
          items: {}
        steps:
          type: array
          minItems: 1
          maxItems: 20
          items: {$ref: '#/components/schemas/AnimationStep'}
...
    ChapterProgress:
      type: object
      required:
        - chapterId
        - chapterNumber
        - title
        - chatCount
        - classroomCount
        - animationCount
        - codeRunCount
        - eventCount
        - totalActivities
      properties:
        chapterId: {type: string}
        chapterNumber: {type: integer}
        title: {type: string}
        chatCount: {type: integer, format: int64}
        classroomCount: {type: integer, format: int64}
        animationCount: {type: integer, format: int64}
        codeRunCount: {type: integer, format: int64}
        eventCount: {type: integer, format: int64}
        totalActivities: {type: integer, format: int64}
        lastActivityAt: {type: string, format: date-time, nullable: true}
    LearningProgress:
      type: object
      required: [totalActivities, chapters]
      properties:
        totalActivities: {type: integer, format: int64}
        chapters:
          type: array
          items: {$ref: '#/components/schemas/ChapterProgress'}
    PersistedWorkbenchScene:
      type: object
      required: [recordId, chapterId, definition]
      properties:
        recordId: {type: string}
        chapterId: {type: string}
        definition: {$ref: '#/components/schemas/AnimationDefinition'}
        updatedAt: {type: string, format: date-time, nullable: true}
    LearningWorkbench:
      type: object
      required: [progress]
      properties:
        currentChapterId: {type: string, nullable: true}
        progress: {$ref: '#/components/schemas/LearningProgress'}
        scene:
          allOf:
            - {$ref: '#/components/schemas/PersistedWorkbenchScene'}
          nullable: true
    LearningEventRequest:
      type: object
      required: [eventType]
      properties:
        eventType:
          type: string
          description: Classroom answers, animation observations, and code reviews are recorded only by their domain endpoints.
          enum: [RESOURCE_VIEW, RESOURCE_DOWNLOAD, REVIEW_COMPLETED, WEAKNESS_RECORDED]
        chapterId: {type: string, maxLength: 64, nullable: true}
        referenceId: {type: string, maxLength: 96, nullable: true}
        payload: {type: object, additionalProperties: true, nullable: true}
    LearningEvent:
      type: object
      required: [id, eventType, createdAt]
      properties:
        id: {type: integer, format: int64}
        eventType: {type: string}
        chapterId: {type: string, nullable: true}
        referenceId: {type: string, nullable: true}
        createdAt: {type: string, format: date-time}
...
    SandboxConfigCapability:
      type: object
      required: [available]
      properties:
        available: {type: boolean}
        reason:
          type: string
          nullable: true
          enum: [NOT_CONFIGURED, PERSISTED_CONFIGURATION_DISABLED, SANDBOX_CONFIG_INVALID]
        configuration:
          allOf: [{$ref: '#/components/schemas/SandboxConfig'}]
          nullable: true
        runtime:
          allOf: [{$ref: '#/components/schemas/SandboxRuntimeStatus'}]
          nullable: true
    SandboxRuntimeStatus:
      type: object
      required: [codeExecutionConfigured, source, provider]
      properties:
        codeExecutionConfigured: {type: boolean}
        source: {type: string, enum: [persisted, environment, unknown]}
        provider: {type: string, enum: [PISTON, JUDGE0, UNKNOWN]}
        checkedAt: {type: string, format: date-time, nullable: true}
    SandboxConfig:
      type: object
      required: [provider, baseUrl, enabled, updatedAt]
      properties:
        provider: {type: string, enum: [PISTON, JUDGE0]}
        baseUrl: {type: string, format: uri, maxLength: 2048}
        enabled: {type: boolean}
        lastConnectionTestStatus:
          type: string
          nullable: true
          description: Diagnostic result of the latest explicit probe.
        lastConnectionTestedAt: {type: string, format: date-time, nullable: true}
        updatedAt: {type: string, format: date-time}
    UpdateSandboxConfigRequest:
      type: object
      required: [provider, baseUrl, enabled]
      properties:
        provider: {type: string, enum: [PISTON, JUDGE0]}
        baseUrl: {type: string, format: uri, minLength: 1, maxLength: 2048}
        enabled: {type: boolean}
    SandboxConnectionTest:
      type: object
      required: [connected, code]
      properties:
        connected: {type: boolean}
        code:
          type: string
          pattern: '^[A-Z0-9_]{1,64}$'
          description: Controlled result code; no upstream response details are exposed.
...